+++ extension/src/components/Address/index.tsx
@@ -1,15 +1,13 @@
+import { shortenAddress, validateAddress } from '@/utils'
 import cn from 'classnames'
 import copy from 'copy-to-clipboard'
 import React from 'react'
 import { RiExternalLinkLine, RiFileCopyLine } from 'react-icons/ri'
 import { EXPLORER_URL } from '../../chains'
-
 import { useRoute } from '../../panel/routes'
-import { validateAddress } from '../../panel/utils'
-import Blockie from '../Blockie'
+import { Blockie } from '../Blockie'
 import Box from '../Box'
 import IconButton from '../IconButton'
-
 import classes from './style.module.css'
 
 interface Props {
@@ -17,16 +15,6 @@ interface Props {
   explorerLink?: boolean
   copyToClipboard?: boolean
   className?: string
-}
-
-const VISIBLE_START = 4
-const VISIBLE_END = 4
-
-export const shortenAddress = (address: string): string => {
-  const checksumAddress = validateAddress(address)
-  const start = checksumAddress.substring(0, VISIBLE_START + 2)
-  const end = checksumAddress.substring(42 - VISIBLE_END, 42)
-  return `${start}...${end}`
 }
 
 const Address: React.FC<Props> = ({
