+++ extension/src/components/Select/index.tsx
@@ -1,8 +1,8 @@
-import * as React from 'react'
-import Select, { Props } from 'react-select'
+import { CSSProperties } from 'react'
+import BaseSelect, { Props } from 'react-select'
 
 export const selectStyles = {
-  control: (provided: React.CSSProperties, state: any) => ({
+  control: (provided: CSSProperties, state: any) => ({
     ...provided,
     fontFamily: "'Roboto Mono', monospace",
     fontSize: '14px',
@@ -15,18 +15,18 @@ export const selectStyles = {
       borderColor: 'white',
     },
   }),
-  valueContainer: (provided: React.CSSProperties) => ({
+  valueContainer: (provided: CSSProperties) => ({
     ...provided,
     padding: '0px 8px',
   }),
-  input: (provided: React.CSSProperties) => ({
+  input: (provided: CSSProperties) => ({
     ...provided,
     color: 'white',
     margin: 0,
     paddingBottom: 0,
     paddingTop: 0,
   }),
-  option: (provided: React.CSSProperties, state: any) => ({
+  option: (provided: CSSProperties, state: any) => ({
     ...provided,
     background: state.isSelected
       ? 'rgba(217, 212, 173, 0.2)'
@@ -39,25 +39,23 @@ export const selectStyles = {
       background: 'rgba(217, 212, 173, 0.1)',
     },
   }),
-  menu: (provided: React.CSSProperties) => ({
+  menu: (provided: CSSProperties) => ({
     ...provided,
     zIndex: 10,
     borderRadius: 0,
     background: 'rgb(0 0 0 / 95%)',
     marginTop: 0,
   }),
-  menuList: (provided: React.CSSProperties) => ({
+  menuList: (provided: CSSProperties) => ({
     ...provided,
     padding: 0,
   }),
-  singleValue: (provided: React.CSSProperties) => ({
+  singleValue: (provided: CSSProperties) => ({
     ...provided,
     color: 'white',
   }),
 }
 
-const StyledSelect: React.FC<Props> = (props) => {
-  return <Select {...props} styles={selectStyles as any} />
-}
-
-export default StyledSelect
+export const Select = (props: Props) => (
+  <BaseSelect {...props} styles={selectStyles as any} />
+)
