+++ extension/src/components/index.ts
@@ -1,15 +1,19 @@
 export { default as Address } from './Address'
 export { Alert } from './Alert'
 export { default as BlockButton } from './BlockButton'
+export { Blockie } from './Blockie'
 export { default as BlockLink } from './BlockLink'
 export { default as Box } from './Box'
 export { default as BoxButton } from './BoxButton'
 export { default as Button } from './Button'
+export { ConfirmationModal, useConfirmationModal } from './ConfirmationModal'
 export { default as ConnectionStack } from './ConnectionStack'
 export { default as Field } from './Field'
 export { default as Flex } from './Flex'
 export { default as IconButton } from './IconButton'
-export { default as Select } from './Select'
+export { RouteBubble } from './RouteBubble'
+export { Select, selectStyles } from './Select'
 export { default as Spinner } from './Spinner'
 export { default as Tag } from './Tag'
+export { ZodiacToastContainer, toastClasses } from './Toast'
 export { default as ToggleButton } from './ToggleButton'
