+++ extension/src/panel/routes/ConnectButton/index.tsx
@@ -1,3 +1,5 @@
+import { Alert, Button, Flex, Tag } from '@/components'
+import { shortenAddress, validateAddress } from '@/utils'
 import { invariant } from '@epic-web/invariant'
 import classNames from 'classnames'
 import { ZeroAddress } from 'ethers'
@@ -5,11 +7,8 @@ import React from 'react'
 import { RiAlertLine } from 'react-icons/ri'
 import { ChainId, parsePrefixedAddress } from 'ser-kit'
 import { CHAIN_NAME } from '../../../chains'
-import { Alert, Button, Flex, Tag } from '../../../components'
-import { shortenAddress } from '../../../components/Address'
 import { ProviderType, Route } from '../../../types'
 import { useInjectedWallet, useWalletConnect } from '../../providers'
-import { validateAddress } from '../../utils'
 import { isConnectedTo } from '../routeHooks'
 import metamaskLogoUrl from './metamask-logo.svg'
 import classes from './style.module.css'
