+++ extension/src/panel/transactions/SimulationStatus.tsx
@@ -1,6 +1,6 @@
+import { Flex, Spinner, Tag } from '@/components'
 import React from 'react'
 import { RiExternalLinkLine, RiGitBranchLine } from 'react-icons/ri'
-import { Flex, Spinner, Tag } from '../../components'
 import { useProvider } from '../providers/ProvideProvider'
 import { TransactionState } from '../state'
 import { ExecutionStatus } from '../state/reducer'
