+++ extension/src/panel/transactions/Translate.tsx
@@ -1,7 +1,6 @@
+import { IconButton } from '@/components'
 import React from 'react'
 import { BiWrench } from 'react-icons/bi'
-
-import { IconButton } from '../../components'
 import { ForkProvider } from '../providers'
 import { useProvider } from '../providers/ProvideProvider'
 import { TransactionState, useDispatch, useTransactions } from '../state'
