+++ extension/src/panel/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './addressValidation'
 export * from './decodeError'
 export * from './roleKey'
