+++ extension/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './addressValidation'
+export { shortenAddress } from './shortenAddress'
